+++ src/main/java/nhom2/NganHang.java
@@ -19,8 +19,8 @@ public class NganHang {
         khoanGui.add(soNganHang);
     }
 
-    public int getLais(Date today) {
-        int ans = 0;
+    public long getLais(Date today) {
+        long ans = 0;
         for (int i = 0; i < khoanGui.size(); ++i) {
             if (khoanGui.get(i).getNgayTra().isEqual(today)) {
                 ans += khoanGui.get(i).getLai();
